Projectile particles fade out
projectiles now fade out rather than deleting when projectile itself stops
